Rename README.md to New Data Team Creation
While working at Excel Clinical Research, I have given the opportunity to create, to train, to implement, and to lead a new Data Management Team which was first time to be done at Excel Clinical Research due to the need for it at a such growing clinical research site at that time. From March 2022 till April 2023.